requirements.txt

---

# 📦 `requirements.txt`

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# LLM & embeddings
openai>=1.3.0
cohere>=4.40.0

# Vector database client (pgvector)
supabase>=2.3.0

# PDF generation & Arabic RTL support
reportlab>=4.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Environment variables
python-dotenv>=1.0.1
